LayoutTests/fast/custom-elements/connected-callbacks.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: connectedCallback</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="connectedCallback must be enqueued whenever custom element is inserted into a document">
<link rel="help" href="https://w3c.github.io/webcomponents/spec/custom/#dfn-connected-callback">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

var calls = [];
class MyCustomElement extends HTMLElement {
    connectedCallback() { calls.push('connected', this); }
    disconnectedCallback() { calls.push('disconnected', this); }
}
customElements.define('my-custom-element', MyCustomElement);

test(function () {
    var instance = document.createElement('my-custom-element');

    calls = [];
    document.body.appendChild(instance);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting a custom element into a document must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');

    calls = [];
    var parent = document.createElement('div');
    parent.appendChild(instance);
    assert_array_equals(calls, []);
}, 'Inserting a custom element into a detached node must not enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var documentWithoutWindow = document.implementation.createHTMLDocument();

    calls = [];
    documentWithoutWindow.body.appendChild(instance);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting a custom element into a window-less document must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var parent = document.createElement('div');
    parent.appendChild(instance);

    calls = [];
    document.body.appendChild(parent);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting an ancestor of a custom element into a document must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var parent = document.createElement('div');
    parent.appendChild(instance);
    var documentWithoutWindow = document.implementation.createHTMLDocument();

    calls = [];
    documentWithoutWindow.body.appendChild(parent);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting an ancestor of custom element into a window-less document must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    var shadowRoot = host.attachShadow({mode: 'closed'});
    document.body.appendChild(host);

    calls = [];
    shadowRoot.appendChild(instance);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting a custom element into a connected shadow tree must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    var shadowRoot = host.attachShadow({mode: 'closed'});
    shadowRoot.appendChild(instance);

    calls = [];
    document.body.appendChild(host);
    assert_array_equals(calls, ['connected', instance]);
}, 'Inserting the shadow host of a shadow tree with a custom element into a document must enqueue and invoke connectedCallback');

test(function () {
    var instance = document.createElement('my-custom-element');
    var host = document.createElement('div');
    var shadowRoot = host.attachShadow({mode: 'closed'});

    calls = [];
    shadowRoot.appendChild(instance);
    assert_array_equals(calls, []);
}, 'Inserting a custom element into a detached shadow tree must not enqueue and invoke connectedCallback');

</script>
</body>
</html>
